requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.16.18
botocore==1.19.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-multiselectfield==0.1.12
django-rest-knox==4.1.0
djangorestframework==3.12.1
dynaconf==3.1.2
hvac==0.10.5
idna==2.10
isort==5.6.4
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.11
lazy-object-proxy==1.4.3
loguru==0.5.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
validators==0.18.1
whitenoise==5.2.0
win32-setctime==1.0.3
wrapt==1.12.1
